% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gsfile.R
\name{download_gsfile}
\alias{download_gsfile}
\title{Download a remote file on Google Storage}
\usage{
download_gsfile(remote_path, dest_dir)
}
\arguments{
\item{remote_path}{a remote path on Google Storage}

\item{dest_dir}{a destination directory to download}
}
\description{
Download a remote file on Google Storage
}
