% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_wrapper.R
\name{fread_wrapper}
\alias{fread_wrapper}
\title{Wrap data.table::fread for bgz or pipe}
\usage{
fread_wrapper(path, extra_pipe_cmd = NULL, ...)
}
\arguments{
\item{path}{a path to file}

\item{extra_pipe_cmd}{if specified, this will be added to pipe}

\item{...}{extra parameters for data.table::fread}
}
\value{
a data.frame
}
\description{
Wrap data.table::fread for bgz or pipe
}
