% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwrite_wrapper.R
\name{fwrite_wrapper}
\alias{fwrite_wrapper}
\title{Wrap data.table::fwrite for bgz or pipe}
\usage{
fwrite_wrapper(x, path, sep = "\\t", na = "NA", ...)
}
\arguments{
\item{x}{a data.frame to write}

\item{path}{a path to file}

\item{sep}{an output separator}

\item{...}{extra parameters for data.table::fwrite}
}
\description{
Wrap data.table::fwrite for bgz or pipe
}
