\name{gs.getacl}
\alias{gs.getacl}
\title{ Function to get the ACL of remote objects }
\description{
    This function allows the user to get the ACL of remote objects.
}
\usage{
gs.getacl(uri)
}
\arguments{
  \item{uri}{ uri of the object to get the ACLs for}
}
\details{
    This function allows the user to browse object metadata concerning access rights. 
    Most features are unimplemented. 
}
\value{
    A XML string.
}
\references{
}
\author{ Carlos J. Gil Bellosta }
\examples{
}
\keyword{ manip }
