% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gsfile.R
\name{read_gsfile}
\alias{read_gsfile}
\title{Read a remote file on Google Storage}
\usage{
read_gsfile(remote_path, extra_pipe_cmd = NULL, cache.dir = "/tmp", ...)
}
\arguments{
\item{remote_path}{a remote path on Google Storage}

\item{extra_pipe_cmd}{if specified, this will be passed to pipe}

\item{cache.dir}{a directory for cache}

\item{...}{extra parameters for data.table::fread}
}
\value{
a data.frame
}
\description{
Read a remote file on Google Storage
}
