% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_gsfile.R
\name{upload_gsfile}
\alias{upload_gsfile}
\title{Upload a local file on Google Storage}
\usage{
upload_gsfile(local_path, remote_path)
}
\arguments{
\item{local_path}{a local path}

\item{remote_path}{a remote path on Google Storage}
}
\description{
Upload a local file on Google Storage
}
