% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsfile.R
\name{write_gsfile}
\alias{write_gsfile}
\title{Write a remote file on Google Storage}
\usage{
write_gsfile(
  x,
  remote_path,
  sep = "\\t",
  overwrite = FALSE,
  cache.dir = "/tmp",
  ...
)
}
\arguments{
\item{x}{a data.frame to write}

\item{remote_path}{a remote path on Google Storage}

\item{sep}{an output separator}

\item{overwrite}{whether to overwrite an existing remote file}

\item{cache.dir}{a directory for cache}

\item{...}{extra parameters for data.table::fread}
}
\description{
Write a remote file on Google Storage
}
