% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_update.R
\name{check_update}
\alias{check_update}
\title{Check whether a remote file is updated}
\usage{
check_update(remote_path, local_path)
}
\arguments{
\item{remote_path}{a remote path on Google Storage}

\item{local_path}{a local path to check whether updated}
}
\value{
whether the file is updated
}
\description{
Check whether a remote file is updated
}
