% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsfile_exists.R
\name{gsfile_exists}
\alias{gsfile_exists}
\title{Check whether a remote file exists}
\usage{
gsfile_exists(remote_path)
}
\arguments{
\item{remote_path}{a remote path on Google Storage}
}
\value{
whether the file exists
}
\description{
Check whether a remote file exists
}
