% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gsfile.R
\name{list_gsfile}
\alias{list_gsfile}
\title{List files on a remote path}
\usage{
list_gsfile(remote_path)
}
\arguments{
\item{remote_path}{a remote path on Google Storage}
}
\value{
a list of files
}
\description{
List files on a remote path
}
